backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
python-multipart==0.0.12
google-generativeai==0.8.3
pydantic==2.10.1
